renamed to imaging_utils and type-checked drift
